chore: Change dev environment url binding
This allows devices on the local network to access the app.

The only caveat is that the default URL `$ blazor watch` etc accesses isn't correct, but the app is accessible at `localhost:5209`.

The above suggests there is a better way to accomplish this behavior, but this was the simplest solution I found without using additional software etc.