refactor: Add generic type to CardPoolComponent and CardComponent
This is a much cleaner (and perhaps more performant) way to restrict CardPoolComponents to aggregating a particular subclass of ICard.

This also moves their code to code-behind files, which is necessary to assign a generic type AFAIK.

CardComponent uses a cascading type parameter to provide its children CardComponents with the matching type.