refactor: Move game initialization code to component constructor
Fluxor state doesn't seem to be available during the constructor phase, so Fluxor-related code (and code it depends on) is staying in OnInitialized.

The difference isn't immediately apparent, but it seems like better practice pending more refactors, given that this game logic code should be separate from the presentation layer anyway.